\name{envOrNULL-class}
\Rdversion{1.1}
\docType{class}
\alias{envOrNULL-class}

\title{Class \code{"envOrNULL"}}
\description{
  A union class used to abstract whether data is in memory or out.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "envOrNULL" in the signature.
}
\examples{
showClass("envOrNULL")
}
\keyword{classes}
