\name{writeH5Data-methods}
\alias{writeH5Data-methods}
\alias{writeH5Data}
\alias{writeH5Data,H5Dataset-method}
\title{Writes H5 Data to an Object}
\description{
  Writes data to an H5Dataset.
}
\usage{
  \S4method{writeH5Data}{H5Dataset}(h5Obj, data, offsets, extents)
}
\arguments{
  \item{h5Obj}{
    The object 
  }
  \item{data}{
    data is the data to be written.
  }
  \item{offsets}{
    data is the data to be written.
  }
  \item{extents}{
    data is the data to be written.
  }
}
\section{Methods}{
\describe{

\item{\code{signature(h5Obj = "H5Dataset")}}{
}
}}
\keyword{methods}
